metrics: report attachment duration with sub-millisecond precision
The duration was divided by time.Millisecond as an integer before being
converted to a float, so it was truncated to whole milliseconds: every
attachment faster than 1ms reported 0. Three of the five testnet sequencers read
a flat 0 while the other two read 2-5ms, which looks like a machine difference
rather than a rounding boundary and cost an investigation before it was
recognised as one.

The metric keeps its name and unit; only the precision changes.

+++ global/global.go
@@ -738,7 +738,9 @@ func (l *Global) registerMetrics() {
 
 func (l *Global) AttachmentFinished(started time.Time, cost int) {
 	l.attachmentsCounter.Inc()
-	l.attachmentTimeMilliseconds.Set(float64(time.Since(started) / time.Millisecond))
+	// divide in float: integer Duration division truncates to whole milliseconds, which reported
+	// every attachment faster than 1ms as 0 and made the metric useless exactly where it matters
+	l.attachmentTimeMilliseconds.Set(float64(time.Since(started)) / float64(time.Millisecond))
 	l.attachmentCostCounter.Add(float64(cost))
 }
